requirements.txt
appdirs==1.4.3
asgiref==3.2.7
astroid==2.4.1
certifi==2020.4.5.1
chardet==3.0.4
contextlib2==0.6.0.post1
cycler==0.10.0
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.6
django-cors-headers==3.3.0
django-crispy-forms==1.9.1
django-dajax==0.9.2
django-dajaxice==0.7
django-embed-video==1.3.3
djangorestframework==3.11.0
docopt==0.6.2
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
isort==4.3.21
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.2.1
mccabe==0.6.1
numpy==1.18.5
Pillow==7.1.2
pipenv==2018.11.26
psycopg2==2.8.6
psycopg2-binary==2.8.6
pylint==2.5.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.3
python-dotenv==0.14.0
pythonanywhere==0.9.4
pytz==2020.1
requests==2.24.0
schema==0.7.2
six==1.14.0
sqlparse==0.3.1
tabulate==0.8.7
toml==0.10.0
urllib3==1.25.10
virtualenv==20.0.20
virtualenv-clone==0.5.4
whitenoise==5.2.0
wrapt==1.12.1


django-heroku
